move queries that were in mutations to queries file
